@@ -13,25 +13,25 @@
   "devDependencies": {
     "@iconify/json": "^2.2.151",
     "@sveltejs/adapter-netlify": "^4.1.0",
-    "@sveltejs/vite-plugin-svelte": "^3.0.0",
+    "@sveltejs/vite-plugin-svelte": "^4.0.0",
     "eslint": "^8.55.0",
     "eslint-config-prettier": "^9.1.0",
     "prettier": "3.2.5",
-    "prettier-plugin-svelte": "^3.2.1",
+    "prettier-plugin-svelte": "^3.2.6",
     "sass": "~1.64.2",
-    "svelte": "^4.2.12",
+    "svelte": "^5.0.0",
     "unplugin-icons": "^0.18.5",
-    "vite": "^5.2.8",
+    "vite": "^5.4.4",
     "vite-imagetools": "^6.2.9"
   },
   "dependencies": {
     "@fontsource/ibm-plex-sans": "^5.0.17",
     "@popperjs/core": "^2.11.8",
-    "@sveltejs/kit": "^2.0.0",
+    "@sveltejs/kit": "^2.5.27",
     "animate.css": "^4.1.1",
     "bootstrap": "~5.2.3",
     "sharp": "0.33.2",
-    "svelte-preprocess": "^5.1.3",
+    "svelte-preprocess": "^6.0.0",
     "sveltekit-i18n": "^2.4.2"
   },
   "packageManager": "yarn@4.1.1"
