@@ -8,7 +8,7 @@
     <div class="col-md-3">
       <h2 class="section-heading">
         {$t("career.title")}
-        <span class="section-border" />
+        <span class="section-border"></span>
       </h2>
       <p class="section-subtitle">{$t("career.subtitle")}</p>
     </div>
@@ -17,39 +17,63 @@
         <!-- Studies -->
         <div class="order-1 order-md-0">
           <ExperienceCard type="school" from="2022" to={$t("career.today")}>
-            <span slot="title">{$t("career.master")}</span>
-            <span slot="description">{$t("career.master.description")}</span>
+            {#snippet title()}
+                        <span >{$t("career.master")}</span>
+                      {/snippet}
+            {#snippet description()}
+                        <span >{$t("career.master.description")}</span>
+                      {/snippet}
           </ExperienceCard>
         </div>
         <div class="order-2 order-md-1">
           <ExperienceCard type="school" from="2021" to="2022">
-            <span slot="title">{$t("career.bachelor")}</span>
-            <span slot="description">{$t("career.bachelor.description")}</span>
+            {#snippet title()}
+                        <span >{$t("career.bachelor")}</span>
+                      {/snippet}
+            {#snippet description()}
+                        <span >{$t("career.bachelor.description")}</span>
+                      {/snippet}
           </ExperienceCard>
         </div>
         <div class="order-3 order-md-2">
           <ExperienceCard type="school" from="2019" to="2021">
-            <span slot="title">{$t("career.technical")}</span>
-            <span slot="description">{$t("career.technical.description")}</span>
+            {#snippet title()}
+                        <span >{$t("career.technical")}</span>
+                      {/snippet}
+            {#snippet description()}
+                        <span >{$t("career.technical.description")}</span>
+                      {/snippet}
           </ExperienceCard>
         </div>
         <!-- Work experiences -->
         <div class="order-0 order-md-0">
           <ExperienceCard type="work" from="2022" to={$t("career.today")}>
-            <span slot="title">{$t("career.covea")}</span>
-            <span slot="description">{$t("career.covea.description")}</span>
+            {#snippet title()}
+                        <span >{$t("career.covea")}</span>
+                      {/snippet}
+            {#snippet description()}
+                        <span >{$t("career.covea.description")}</span>
+                      {/snippet}
           </ExperienceCard>
         </div>
         <div class="order-1 order-md-1">
           <ExperienceCard type="work" from="2021" to="2022">
-            <span slot="title">{$t("career.cgi")}</span>
-            <span slot="description">{$t("career.cgi.description")}</span>
+            {#snippet title()}
+                        <span >{$t("career.cgi")}</span>
+                      {/snippet}
+            {#snippet description()}
+                        <span >{$t("career.cgi.description")}</span>
+                      {/snippet}
           </ExperienceCard>
         </div>
         <div class="order-2 order-md-2">
           <ExperienceCard type="work" from="2020" to="2021">
-            <span slot="title">{$t("career.sdis")}</span>
-            <span slot="description">{$t("career.sdis.description")}</span>
+            {#snippet title()}
+                        <span >{$t("career.sdis")}</span>
+                      {/snippet}
+            {#snippet description()}
+                        <span >{$t("career.sdis.description")}</span>
+                      {/snippet}
           </ExperienceCard>
         </div>
       </div>
